Create models for relations
