+++ api/src/models/favoritesCompModel.js
@@ -1,0 +1,12 @@
+const sequelize = require('../database');
+const { DataTypes } = require('sequelize');
+
+const FavoritesComp = sequelize.define('FavoritesComp', {
+    id: {
+        type: DataTypes.INTEGER,
+        autoIncrement: true,
+        primaryKey: true,
+    },
+});
+
+module.exports = FavoritesComp;
