+++ api/src/models/relations.js
@@ -3,8 +3,9 @@ const Postulant = require('./postulantModel');
 const Admin = require('./adminModel');
 const Company = require('./companyModel');
 const Offers = require('./offersModel');
-const sequelize = require('../database');
-const { DataTypes } = require('sequelize');
+const Aplications = require('./applicationModel');
+const SaveOffer = require('./saveOfferModel');
+const FavoritesComp = require('./favoritesCompModel');
 
 Postulant.belongsTo(User, { foreignKey: 'userId', onDelete: 'CASCADE' });
 User.hasMany(Postulant, { foreignKey: 'userId' });
@@ -21,45 +22,21 @@ User.hasMany(Offers, { foreignKey: 'userId' });
 Offers.belongsTo(Company, { foreignKey: 'idRecruiterOfferCreate', onDelete: 'CASCADE' });
 Company.hasMany(Offers, { foreignKey: 'idRecruiterOfferCreate' });
 
-const Aplications = sequelize.define('Aplications', {
-    id: {
-        type: DataTypes.INTEGER,
-        autoIncrement: true,
-        primaryKey: true,
-    },
-    status:{
-        type: DataTypes.ENUM('send', 'viewed', 'no_select', 'select')
-    }
-});
-
+// Aplicaciones
 Aplications.belongsTo(User, { foreignKey: 'userId', onDelete: 'CASCADE' });
 User.hasMany(Aplications, { foreignKey: 'userId', onDelete: 'CASCADE' });
 
 Aplications.belongsTo(Offers, { foreignKey: 'offerId', onDelete: 'CASCADE' });
 Offers.hasMany(Aplications, { foreignKey: 'offerId', onDelete: 'CASCADE' });
 
-const SaveOffer = sequelize.define('SaveOffer', {
-    id: {
-        type: DataTypes.INTEGER,
-        autoIncrement: true,
-        primaryKey: true,
-    },
-});
-
+// Guardar ofertas
 SaveOffer.belongsTo(User, { foreignKey: 'userId', onDelete: 'CASCADE' });
 User.hasMany(SaveOffer, { foreignKey: 'userId', onDelete: 'CASCADE' });
 
 SaveOffer.belongsTo(Offers, { foreignKey: 'offerId', onDelete: 'CASCADE' });
 Offers.hasMany(SaveOffer, { foreignKey: 'offerId', onDelete: 'CASCADE' });
 
-const FavoritesComp = sequelize.define('FavoritesComp', {
-    id: {
-        type: DataTypes.INTEGER,
-        autoIncrement: true,
-        primaryKey: true,
-    },
-});
-
+// Empresas favoritas 
 FavoritesComp.belongsTo(User, { foreignKey: 'userId', onDelete: 'CASCADE' });
 User.hasMany(FavoritesComp, { foreignKey: 'userId', onDelete: 'CASCADE' });
